refactor: pull request list features
Refactor all the code related to the PR list view features to split the
functionality related to linkifying target branch namesand adding the
PR's source branch name.

Also, the ignore whitespace by default feature was involved in the
refactor and a bug was solved as a consequence, where only the PRs that
were initially loaded with the page were touched, while any PR added
dynamically to the DOM were not affected.

Related to issue #135 and pr #136.

Closes issue #138.