fix: word diffs no longer remove syntax highlight (#158)
When word diffing is enabled for file diffs (which is the default),
Bitbucket asynchronously re-renders each source code line to insert
`<ins>` and `<del>` tags that will highlight with brighter green or red
the changes on that specific line. Since the extension does the syntax
highlighting pass as soon as the file diff loads into the DOM, the word
diffing re-render applied later removes all the manipulation we had
previously done, effectively removing the syntax highlighting and all
color in the line.

The solution is to observe the file diff for this re-render and when it
happens, run the syntax highlighting logic again after the fact in the
specific lines that were changed. However, this is just half of the
solution, because PrismJS by default strips all original markup from the
line, so this would result in a syntax highlighted line, but with no
word diffs. To amend this, the "Keep Markup" plugin has been added so
that the `<ins>` and `<del>` tags are preserved.

Closes #63.