fix(diff): extra leading pluses/minuses in diff (#177)
Bitbucket re-renders diff lines into the DOM when including word diffs
without actually adding any <ins> or <del> tags to their markup. Since
the previous implementation was relying on the presence of those tags to
determine which lines were re-rendered, lines that didn't have them were
not being picked up by our function in the second pass, and those were
the ones that remained with the extra "-" or "+".

Now the implementation has been changed so that we determine which lines
were re-rendered by comparing the contents of those lines before the
word diff has been applied, and after. This solution is not only better
(possibly fail-proof), but also doesn't depend on jQuery, and the
resulting code is smaller, cleaner and more succinct.

Closes #168
Closes #175