feat: add "Show comments" checkbox (#196)
Adds a "Comments" checkbox at the header of the diff that will be checked by default. When unchecked, all the comments in the current diff will be hidden, and will reappear when checked again.

The presence of the checkbox is dynamic. If the current diff has no comments, there will be no checkbox, but if a comment is added, it will show up. It will also be removed if all the comments in the diff are deleted.

Closes #172