Fix bad side-by-side scrolling performance (#233)
Every time there was a scroll event when the side-by-side pane was open,
all lines that had word-diffs in them (`<ins>` and `<del>` tags) would
get syntax-highlighted again, increasing the size of the DOM tree
unnecesarily and infinitely for those nodes. The more the user scrolled,
the bigger the memory and CPU consumption. This caused a tremendous
performance impact.

The logic of the syntax highlighting function was updated so that it adds
a boolean property to the first child of the source code line `<pre>`
DOM node, this property is then going to be used on each subsequent call
to determine if the given line should be highlighted again. It is done
this way because it is the contents of that `<pre>` tag that are mutated
when word diffing occurs, not the wrapping line itself.

Closes #183