feat: Merge commit message template (#243)
Closes #97 

Also see #106 and #243.

Similar to how the pull request template feature works (https://git.io/fAlck), 
it is now possible to also point to a `MESSAGE_COMMIT_TEMPLATE` file
that will override Bitbucket's default commit message for all merging
strategies. The file should be checked in to source control in the
default branch and located in either:

```
/MERGE_COMMIT_TEMPLATE
/docs/MERGE_COMMIT_TEMPLATE
/.github/MERGE_COMMIT_TEMPLATE
/.bitbucket/MERGE_COMMIT_TEMPLATE
```

Again, the filename _IS_ case sensitive, and should not have any
extension (not `.md`, `.txt`, etc.). The file to be used will be the
first to be found in one of these locations from top to bottom, or the
one hosted externally and specified by a URL in the options page.
Externally hosted template files have only been tested with RAW 
GitHub's Gists, and will only work on Chrome, not on Firefox. 
For example: https://git.io/fAlct.
You can use this alternative method if you don't want to include any new
file in your repository.

In addition, this template can contain 'environment variables' that
the extension will care to inject for you, making the message truly
dynamic. The variables supported so far are:

- `id`
- `title`
- `description`
- `sourceBranch`
- `targetBranch`
- `approvedByList`. Note that this will only include the user's display
name and it's username, not its email (see the example below).

For example, for a template file with the following contents...

```
{title} (#{id})

{sourceBranch} => {targetBranch}

{description}

{approvedByList}
```

The resulting commit message can look like this:

```
Homogenize config key names (#612)

homogenize-configs => develop

For consistency’s sake

Approved By: Ronald Rey <reyronald>
```