chore: Upgrade almost every dependency in package.json (#246)
That includes Babel 7 and Webpack 4!

Also, in this pull request I'm removing the Uglyfication/Minification build step performed in production bundles because every time I release a new version, the extension is flagged as suspicious by the Mozilla's Add-on store, which triggers a manual code review request that ultimately ends up in me having to upload the original source code. The e-mail I receive each time reads ...

> This version contains obfuscated, minified, concatenated or otherwise machine-generated code. Please provide the original sources, together with instructions on how to generate the final XPI. Source code must be provided as an archive and uploaded using the source code upload field, which can be done during submission or on the version page in the developer hub. 
>
> Please read through the instructions at https://developer.mozilla.org/en-US/Add-ons/Source_Code_Submission. If sources have not been provided within 7 days, this version will be rejected and removed from our queues. 

This doesn't have any impact for the user since extensions's code, unlike web apps, are downloaded only once and stored locally in their drives when it is installed or updated. So removing this step won't result in a degraded experience.