Add Flow (#247)
Introduce [Flow](http://flow.org/) for static type checking.

There is a lot of type casting through `any` for `.querySelector` and
DOM related operations. This is intentional! The intent is to have the
extension blow up and crash instead of silently failing when Bitbucket's
DOM structure changes and is not as the extension expects, this will
allow me to use those stack traces to pinpoint the exact place of
failure and fix it faster and easier.