Updated New UI Syntax Implementation to Conserve Word Highlighting
The previous implementation included an optimized way of handling DOM updates,
but it resulted in erasing child elements of each line of code.  Bitbucket handles
word comparison highlighting by inserting <ins> and <del> into the code,
therefore these were lost.

The current implementation now highlights each row of code in-place on the DOM, taking a
small performance decrease but allowing the Prism library to ensure that <ins> and <del>
are respected and word highlighting is not lost.